requirements.txt
# Plant Disease Detection AI - Backend Dependencies
# For server_ai_takeover.py

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ML & Image Processing
tensorflow>=2.14.0
numpy>=1.24.0
Pillow>=10.0.0

# HTTP & API
httpx>=0.24.0
python-multipart>=0.0.6

# Generative AI (Gemini)
google-generativeai>=0.3.0

# Environment & Configuration
python-dotenv>=1.0.0

# Optional: Development Tools
# pytest>=7.4.0
# black>=23.10.0
